+++ src/layouts/ExtensionBrowserAction/components/CxFooter/CxFooter.vue
@@ -105,6 +105,10 @@ export default {
         {
           to: 'https://www.medium.com/@myetherwallet',
           class: 'fa-medium'
+        },
+        {
+          to: 'https://t.me/myetherwalletofficial',
+          class: 'fa-telegram'
         }
       ]
     };